Where I Stand
I am getting a null error for the x and y position. Doesn't come up as
syntax, but when I run the project it pops up. I tried to make random
color. I created a color class, but didn't know how to incorporate it
into I think a constructor. And the ball variable is still unknown. I
tried making it ballList.draw(); and BoxBall.draw(); but those don't
work because I don't know how to make the BoxBall draw.